refactor(board): co-locate board-sets store + hook with their UI
Move board-sets-store.ts and use-board-sets.ts from storage/ into
board-sets/, so board-sets is a self-contained slice (UI + state) and
storage/ is a pure shared-persistence layer (boards-db, board-hydration).
