+++ src/features/board/index.ts
@@ -1,6 +1,8 @@
+export { deleteBoardSet, getBoardSets } from "./board-sets/board-sets-store";
 export { BoardSetDeleteDialog } from "./board-sets/delete-dialog";
 export { BoardSetInfoDialog } from "./board-sets/info-dialog";
 export { BoardSetList } from "./board-sets/list";
+export { useBoardSets } from "./board-sets/use-board-sets";
 export { BoardViewer } from "./board-viewer";
 export { importBoardFromUrl } from "./import/import-from-url";
 export { useImportBoardFiles } from "./import/use-import-board-files";
@@ -10,9 +12,7 @@ export {
   boardPath,
 } from "./navigation/board-paths";
 export { BoardNotFoundError, hydrateBoard } from "./storage/board-hydration";
-export { deleteBoardSet, getBoardSets } from "./storage/board-sets-store";
 export { getBoardSet } from "./storage/boards-db";
-export { useBoardSets } from "./storage/use-board-sets";
 export { resolveTranslatedBoard } from "./translation/resolve-translated-board";
 
 export type { BoardRouteParams } from "./navigation/use-board-navigation";
