+++ src/features/board/navigation/use-board-navigation.ts
@@ -1,5 +1,5 @@
 import { useLocation, useNavigate, useParams } from "react-router";
-import { useBoardSets } from "../storage/use-board-sets";
+import { useBoardSets } from "../board-sets/use-board-sets";
 import { boardPath } from "./board-paths";
 
 export interface BoardRouteParams {
